pancake-order-system/order_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
protobuf==6.31.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
temporalio==1.11.1
types-protobuf==6.30.2.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
langchain>=0.3.25
inflect
